Use a stacked widget instead of showing and hiding vertically (linux)
This a part of my efforts of bringing the notifications in the line.

At this moment, the different views in the main window are laid out in one single layout and different parts (like TimeEntryList and TimeEntryEdit) get shown/hidden depending on the backend requests. This was not very easy to maintain because with the addition of more views, the programmer would have to check all the other views and make them hide when his new view goes into place.

QStackedWidget takes care of this as it's a widget wrapper that can contain multiple widgets and display only a single one of them at a time as requested. That means now with the addition of a new view, one would have to only request the QStackedWidget to display his new view and the currently shown one gets hidden without any extra effort.

As a nice bonus, everything seems much faster and more fluid when switching between views