Make all the vectors in related_data protected by mutexes (lib)
Everything is now returned as a std::pair through getters

First element of that pair is a lock that gets closed automatically when scope ends or explicitely by using pair.first.unlock()

Second element is the actual pointer to the vector itself. Everything is now accessed through pointers - this will require some changes to the other methods, too, to increase efectiveness because in a lot of places, the vectors were passed by value.

Closes #2747