Handle adding tags inside the app (linux)
A hacky, workaround-y solution that stores the newly created tags until they actually come from the server side.

This is to prevent them being overwritten (and flashing) inside the app, because after creating a new tag, the library or the backend sends multiple updates with old ones before they start to contain the newly created one.

Closes #1973