Remove almost every redundant Toggl*View class (lib)
These classes are redundant because toggl::view::* classes do exactly the same thing and the Toggl*View classes are an unnecessary layer of reallocations and error-prone code.

Now, this doesn't work, yet, but it compiles. The next step is to restore functionality with this code so we can actually use these simplified methods in real code.