Fix bug with in-app message not showing on app start (mac)
In-app message is always added to the view hierarchy before Main VC.
This results in a situation when main VC is covering in-app message view
and users do not see it.
This fix make Main VC always a root view - adds it to the view hierarchy
always below other views.